check for test progress file presence

+++ model.py
@@ -273,6 +273,9 @@ class DistrSystem(object):
 		else:
 			self.initialize_model(session, self.args.load)
 			while (True):
+				if not os.path.exists('test_progress.csv'):
+					time.sleep(self.args.sleep_time)
+					continue
 				if self.args.inst_id == 1:
 					with open('test_progress.csv', 'w') as f:
 						f.write(','*(3*self.args.num_inst-1) + '\n')
